Update to junit5/assert4j
Fixes #127
