+++ dataflow-website/batch-developer-guides/batch/batchsamples/billrun/src/test/java/io/spring/billrun/BillRunApplicationTests.java
@@ -1,42 +1,32 @@
 package io.spring.billrun;
 
-import java.util.List;
-
 import io.spring.billrun.model.Bill;
-import javax.sql.DataSource;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import org.springframework.batch.test.context.SpringBatchTest;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.context.junit4.SpringRunner;
 
-import static org.junit.Assert.assertEquals;
+import javax.sql.DataSource;
+import java.util.List;
 
-@RunWith(SpringRunner.class)
+import static org.assertj.core.api.Assertions.assertThat;
+
 @SpringBootTest
-@SpringBatchTest
-public class BillrunApplicationTests {
+public class BillRunApplicationTests {
 
 	@Autowired
 	private DataSource dataSource;
 
 	private JdbcTemplate jdbcTemplate;
 
-	@Before
-	public void setup()  {
+	@BeforeEach
+	public void setup() {
 		this.jdbcTemplate = new JdbcTemplate(this.dataSource);
 	}
 
 	@Test
 	public void testJobResults() {
-		testResult();
-	}
-
-	private void testResult() {
 		List<Bill> billStatements = this.jdbcTemplate.query("select id, " +
 						"first_name, last_name, minutes, data_usage, bill_amount " +
 						"FROM bill_statements ORDER BY id",
@@ -44,14 +34,15 @@ public class BillrunApplicationTests {
 						rs.getString("FIRST_NAME"), rs.getString("LAST_NAME"),
 						rs.getLong("DATA_USAGE"), rs.getLong("MINUTES"),
 						rs.getDouble("bill_amount")));
-		assertEquals(5, billStatements.size());
 
+		assertThat(billStatements.size()).isEqualTo(5);
 		Bill billStatement = billStatements.get(0);
-		assertEquals(6, billStatement.getBillAmount(), 1e-15);
-		assertEquals("jane", billStatement.getFirstName());
-		assertEquals("doe", billStatement.getLastName());
-		assertEquals(new Long(1), billStatement.getId());
-		assertEquals(new Long(500), billStatement.getMinutes());
-		assertEquals(new Long(1000), billStatement.getDataUsage());
+		assertThat(billStatement.getBillAmount()).isEqualTo(6.0);
+		assertThat(billStatement.getFirstName()).isEqualTo("jane");
+		assertThat(billStatement.getLastName()).isEqualTo("doe");
+		assertThat(billStatement.getId()).isEqualTo(1);
+		assertThat(billStatement.getMinutes()).isEqualTo(500);
+		assertThat(billStatement.getDataUsage()).isEqualTo(1000);
+
 	}
 }
