+++ dataflow-website/batch-developer-guides/batch/batchsamples/billsetuptask/src/test/java/io/spring/billsetuptask/BillSetuptaskApplicationTests.java
@@ -1,0 +1,26 @@
+package io.spring.billsetuptask;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+import javax.sql.DataSource;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest
+class BillSetupTaskApplicationTests {
+
+	@Autowired
+	private DataSource dataSource;
+
+	@Test
+	public void testRepository() {
+		JdbcTemplate jdbcTemplate = new JdbcTemplate(this.dataSource);
+		int result = jdbcTemplate.queryForObject(
+				"SELECT COUNT(*) FROM BILL_STATEMENTS", Integer.class);
+		assertThat(result).isEqualTo(0);
+	}
+
+}
