--- dataflow-website/batch-developer-guides/batch/batchsamples/billsetuptask/src/test/java/io/spring/billsetuptask/BillsetuptaskApplicationTests.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2019 the original author or authors.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *          https://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package io.spring.billsetuptask;
-
-import javax.sql.DataSource;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import static org.junit.Assert.assertEquals;
-
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class BillsetuptaskApplicationTests {
-
-	@Autowired
-	private DataSource dataSource;
-
-	@Test
-	public void testRepository() {
-		JdbcTemplate jdbcTemplate = new JdbcTemplate(this.dataSource);
-		int result = jdbcTemplate.queryForObject(
-				"SELECT COUNT(*) FROM BILL_STATEMENTS", Integer.class);
-
-		assertEquals(0, result);
-	}
-}
